requirements.txt
yfinance>=0.2.31
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
mplfinance>=0.12.10b0
pyarrow>=12.0.0
PyYAML>=6.0.1
pytest>=8.2.0
pytest-cov>=5.0.0
